requirements_llm.txt
# Additional requirements for LLM integration
anthropic>=0.25.0
python-dotenv>=1.0.0
